Add files via upload
Implementation of a 2D PD controller
To test different trajectories, you need to modify the variable trajhandle (uncomment the desired 'trajhandle' variable and comment the undesired ones) inside runsim.m to point to the appropriate function. 
The PD controller was initiated in the file controller.m so that the simulated quadrotor can follow a desired trajectory.